consolidating classifier related modules
