+++ mev_inspect/classifiers/specs/__init__.py
@@ -1,0 +1,4 @@
+from .uniswap import UNISWAP_CLASSIFIER_SPECS
+from .curve import CURVE_CLASSIFIER_SPECS
+
+ALL_CLASSIFIER_SPECS = UNISWAP_CLASSIFIER_SPECS + CURVE_CLASSIFIER_SPECS
