+++ mev_inspect/classifiers/specs/curve.py
@@ -241,4 +241,4 @@ CURVE_LIQUIDITY_GAUGES = [
     ),
 ]
 
-CLASSIFIER_SPECS = [*CURVE_BASE_POOLS, *CURVE_META_POOLS, *CURVE_LIQUIDITY_GAUGES]
+CURVE_CLASSIFIER_SPECS = [*CURVE_BASE_POOLS, *CURVE_META_POOLS, *CURVE_LIQUIDITY_GAUGES]
