+++ mev_inspect/classifiers/specs/uniswap.py
@@ -118,7 +118,7 @@ WETH_SPEC = ClassifierSpec(
 )
 
 
-CLASSIFIER_SPECS = [
+UNISWAP_CLASSIFIER_SPECS = [
     *UNISWAP_V3_CONTRACT_SPECS,
     *UNISWAPPY_V2_CONTRACT_SPECS,
     WETH_SPEC,
